+++ rocketpy/simulation/__init__.py
@@ -1,5 +1,15 @@
 from .flight import Flight
+from .flight_comparator import FlightComparator
 from .flight_data_exporter import FlightDataExporter
 from .flight_data_importer import FlightDataImporter
 from .monte_carlo import MonteCarlo
 from .multivariate_rejection_sampler import MultivariateRejectionSampler
+
+__all__ = [
+    "Flight",
+    "FlightComparator",
+    "FlightDataExporter",
+    "FlightDataImporter",
+    "MonteCarlo",
+    "MultivariateRejectionSampler",
+]